fix: better diffs
